PRQuadTreeTest.java
import student.TestCase;
import java.io.ByteArrayOutputStream;
import java.io.PrintStream;
import java.util.List;

/**
 * Tests the PRQuadTree class to ensure its functionality for managing points
 * in a two-dimensional space, including insertion, removal, and searching
 * capabilities.
 * 
 * @author kinjalpandey, architg03
 * @version 02/25/24
 */
public class PRQuadTreeTest extends TestCase {

    private PRQuadTree quadTree;
    private Point point1;
    private Point point2;
    private Point point3;
    private final ByteArrayOutputStream outContent =
        new ByteArrayOutputStream();
    private final PrintStream originalOut = System.out;

    /**
     * Sets up the test cases by initializing a PRQuadTree and several Points.
     */
    public void setUp() {
        // Initialize the PRQuadTree with a specified boundary size
        quadTree = new PRQuadTree(1024);

        // Initialize points within the boundary
        point1 = new Point("TestPoint1", 100, 100);
        point2 = new Point("TestPoint2", 200, 200);
        point3 = new Point("TestPoint1", 300, 300); // Same name as point1,
                                                    // different coordinates
        System.setOut(new PrintStream(outContent));
    }


    /**
     * Resets the standard output to its original form
     */
    public void tearDown() {
        System.setOut(originalOut);
    }


    /**
     * Tests removing points from the PRQuadTree.
     */
    public void testRemove() {
        // Insert a point and then remove it
        quadTree.insert(point1);
        boolean removed = quadTree.remove(point1);
        assertTrue("Point should be removed successfully", removed);

        // Verify removal by attempting to find the point
        List<Point> foundPoints = quadTree.searchByName(point1.getName());
        assertTrue("No points should be found after removal", foundPoints
            .isEmpty());

        // Attempt to remove a non-existent point
        removed = quadTree.remove(point3);
        assertFalse("Attempt to remove a non-existent point should fail",
            removed);
    }


    /**
     * Test the searchByCoordinates method for a point within a LeafNode.
     */
    public void testSearchLeafNode() {
        Point point = new Point("Point1", 10, 10);
        quadTree.insert(point); // Assuming insert method exists
        Point foundPoint = quadTree.searchByCoordinates(10, 10);
        assertNotNull(foundPoint);
        assertEquals(point, foundPoint);
    }


    /**
     * Test the searchByCoordinates method for a point not within a LeafNode.
     */
    public void testSearchLeafNodeNotFound() {
        Point foundPoint = quadTree.searchByCoordinates(10, 10);
        assertNull(foundPoint);
    }


    /**
     * Test the searchByCoordinates method for a point with replaced equality
     * check with false.
     */
    public void testSearchWithFalseMutation() {
        Point point = new Point("Point1", 50, 50);
        quadTree.insert(point);
        Point foundPoint = quadTree.searchByCoordinates(50, 50);
        assertNotNull(foundPoint); // The point should not be found if the
                                   // equality check is replaced with false
    }


    /**
     * Test the searchByCoordinates method for a point with replaced equality
     * check with true.
     */
    public void testSearchWithTrueMutation() {
        Point point = new Point("Point1", 50, 50);
        quadTree.insert(point);
        Point foundPoint = quadTree.searchByCoordinates(49, 49);
        assertNull(foundPoint); // The point should not be found if the equality
                                // check is replaced with true
    }


    /**
     * Test the searchByCoordinates method for a null or FlyweightNode.
     */
    public void testSearchFlyweightNode() {
        Point foundPoint = quadTree.searchByCoordinates(0, 0);
        assertNull(foundPoint); // Should return null if the node is a
                                // FlyweightNode or null
    }

    /**
     * public void testSearchByCoordinatesMutation() {
     * // Insert points into the quadtree
     * quadTree.insert(point1);
     * quadTree.insert(point2);
     * 
     * // Test equality check mutation
     * Point result = quadTree.searchByCoordinates(100, 100);
     * assertNotNull("Point should be found", result);
     * assertEquals("Point name should match", point1.getName(),
     * result.getName());
     * 
     * // Test arithmetic operation mutation
     * result = quadTree.searchByCoordinates(200, 200);
     * assertNotNull("Point should be found", result);
     * assertEquals("Point name should match", point2.getName(),
     * result.getName());
     * }
     **/

}
